requirements.txt
black==22.6.0
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
tomli==2.0.1
typing_extensions==4.3.0
anyio==3.6.1
beautifulsoup4==4.11.1
black==22.6.0
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
mypy-extensions==0.4.3
lxml==4.9.1
numpy==1.23.1
pandas==1.4.3
pipreqs==0.4.11
psycopg2==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
yarg==0.1.9
